Refactoring: tutorial-component folder renamed to tutorial
