+++ warcommands-app/src/app/tutorial/local-storage-alias.module.ts
@@ -1,6 +1,6 @@
 import { forwardRef, NgModule } from "@angular/core";
-import { LocalStorageGameTutorialRepository } from "../../warcommands/tutorial-component/infrastructure/local-storage/local-storage-game-tutorial-repository";
-import { GameTutorialRepository } from "src/warcommands/tutorial-component/domain/tutorial-component/services/game-tutorial-repository.interface";
+import { LocalStorageGameTutorialRepository } from "../../warcommands/tutorial/infrastructure/local-storage/local-storage-game-tutorial-repository";
+import { GameTutorialRepository } from "src/warcommands/tutorial/domain/tutorial/services/game-tutorial-repository.interface";
 @NgModule({
     providers: [
         { provide: 'GameTutorialRepository', useExisting: forwardRef(() => LocalStorageGameTutorialRepository) }
