+++ warcommands-app/src/app/tutorial/local-storage-provider.module.ts
@@ -1,5 +1,5 @@
 import {NgModule} from "@angular/core";
-import * as LocalStorageGameTutorialRepositoryProvider from "src/warcommands/tutorial-component/infrastructure/angular/factory-providers/tutorial-component/game-tutorial-repository.provider";
+import * as LocalStorageGameTutorialRepositoryProvider from "src/warcommands/tutorial/infrastructure/angular/factory-providers/tutorial-component/game-tutorial-repository.provider";
 
 @NgModule({
     providers: [
