+++ warcommands-app/src/app/tutorial/ngrx-provider-alias.module.ts
@@ -1,6 +1,6 @@
 import { NgModule, forwardRef } from "@angular/core";
-import { NgrxTutorialComponentToggleService } from "src/warcommands/tutorial-component/infrastructure/ngrx/turotial-component/ngrx-tutorial-component-toggle.service";
-import { TutorialComponentToggleServiceInterface } from "src/warcommands/tutorial-component/domain/tutorial-component/services/tutorial-component-toggle-service.interface";
+import { NgrxTutorialComponentToggleService } from "src/warcommands/tutorial/infrastructure/ngrx/turotial-component/ngrx-tutorial-component-toggle.service";
+import { TutorialComponentToggleServiceInterface } from "src/warcommands/tutorial/domain/tutorial/services/tutorial-component-toggle-service.interface";
 
 @NgModule({
     providers: [
