+++ warcommands-app/src/app/tutorial/ngrx-provider.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 
 import * as NgrxTutorialComponentRepositoryProvider
-    from "src/warcommands/tutorial-component/infrastructure/angular/factory-providers/tutorial-component/ngrx-tutorial-component-toggle.provider";
+    from "src/warcommands/tutorial/infrastructure/angular/factory-providers/tutorial-component/ngrx-tutorial-component-toggle.provider";
 
 @NgModule({
     providers: [
