+++ warcommands-app/src/app/tutorial/tutorial-component-provider.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from "@angular/core";
 import { NgrxProviderAliasModule } from "./ngrx-provider-alias.module";
 import { NgrxProviderModule } from "./ngrx-provider.module";
-import * as TutorialOverlayProvider from "src/warcommands/tutorial-component/infrastructure/angular/factory-providers/tutorial-component/tutorial-overlay.provider";
-import * as TutorialComponentProvider from "src/warcommands/tutorial-component/infrastructure/angular/factory-providers/tutorial-component/tutorial-component.provider";
-import * as GameTutorialProvider from "src/warcommands/tutorial-component/infrastructure/angular/factory-providers/tutorial-component/game-tutorial.provider";
-import * as GameTutorialRepositoryProvider from "src/warcommands/tutorial-component/infrastructure/angular/factory-providers/tutorial-component/game-tutorial-repository.provider";
+import * as TutorialOverlayProvider from "src/warcommands/tutorial/infrastructure/angular/factory-providers/tutorial-component/tutorial-overlay.provider";
+import * as TutorialComponentProvider from "src/warcommands/tutorial/infrastructure/angular/factory-providers/tutorial-component/tutorial-component.provider";
+import * as GameTutorialProvider from "src/warcommands/tutorial/infrastructure/angular/factory-providers/tutorial-component/game-tutorial.provider";
+import * as GameTutorialRepositoryProvider from "src/warcommands/tutorial/infrastructure/angular/factory-providers/tutorial-component/game-tutorial-repository.provider";
 
 @NgModule({
     imports: [
