+++ warcommands-app/src/app/tutorial/tutorial/tutorial.component.spec.ts
@@ -3,7 +3,7 @@ import {ComponentFixture, TestBed, waitForAsync} from '@angular/core/testing';
 import {TutorialComponent} from './tutorial.component';
 import {
     TutorialComponentToggleServiceInterface
-} from "../../../warcommands/tutorial-component/domain/tutorial-component/services/tutorial-component-toggle-service.interface";
+} from "../../../warcommands/tutorial/domain/tutorial/services/tutorial-component-toggle-service.interface";
 import {IntroductionComponent} from "../introduction/introduction.component";
 import {MatIconModule} from "@angular/material/icon";
 import {MatTabsModule} from "@angular/material/tabs";
