+++ warcommands-app/src/app/tutorial/tutorial/tutorial.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { TutorialComponentToggleServiceInterface } from "src/warcommands/tutorial-component/domain/tutorial-component/services/tutorial-component-toggle-service.interface";
+import { TutorialComponentToggleServiceInterface } from "src/warcommands/tutorial/domain/tutorial/services/tutorial-component-toggle-service.interface";
 
 @Component({
     selector: 'app-tutorial',
