+++ warcommands-app/src/warcommands/tutorial/domain/tutorial/events/tutorial-first-time-opened.event.ts
