+++ warcommands-app/src/warcommands/tutorial/infrastructure/angular/factory-providers/event-bus/tutorial-event-register.provider.ts
