+++ warcommands-app/src/warcommands/tutorial/infrastructure/event-bus/tutorial-first-time-opened-event-handler.ts
@@ -2,8 +2,8 @@ import { EventHandlerInterface } from "../../../shared/domain/event-bus/event-ha
 import { EventBusInterface } from "../../../shared/domain/event-bus/event-bus-interface";
 import {
     TutorialComponentToggleServiceInterface
-} from "../../domain/tutorial-component/services/tutorial-component-toggle-service.interface";
-import { TutorialEventTypes } from "../../domain/tutorial-component/events/tutorial-event-types.enum";
+} from "../../domain/tutorial/services/tutorial-component-toggle-service.interface";
+import { TutorialEventTypes } from "../../domain/tutorial/events/tutorial-event-types.enum";
 
 export class TutorialFirstTimeOpenedEventHandler implements EventHandlerInterface {
